Impl `inputGuard`, improve config write, code cleanup - ncm: Added `GetDistributionType`

+++ source/install/HttpWorker.cpp
@@ -2,6 +2,7 @@
 #include "util/i18n.hpp"
 #include "nx/error.hpp"
 #include "facade.hpp"
+#include <sstream>
 #include <thread>
 
 namespace app::install
@@ -121,7 +122,7 @@ namespace app::install
 #endif
             app::facade::SendInstallProgress((double)installProgress);
             std::stringstream x;
-            x << (int)(installProgress);
+            x << installProgress;
             app::facade::SendInstallBarText(x.str() + "%");
         }
         std::string ncaIdStr = nx::nca::GetNcaIdString(ncaId);
